[feat] viewpager infomefragmnet
