+++ app/src/main/java/com/example/seminar5/ui/MainActivity.kt
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.seminar5.R
 import com.example.seminar5.databinding.ActivityMainBinding
+import com.example.seminar5.ui.bottom.*
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
