+++ app/src/main/java/com/example/seminar5/ui/bottom/ForyouFragment.kt
@@ -1,4 +1,4 @@
-package com.example.seminar5.ui
+package com.example.seminar5.ui.bottom
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,19 +6,19 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.seminar5.R
-import com.example.seminar5.databinding.FragmentSearchBinding
+import com.example.seminar5.databinding.FragmentForyouBinding
 
 
-class SearchFragment : Fragment() {
-    private var _binding : FragmentSearchBinding ?= null
-    private val binding get() = _binding ?: error("뷰 참조를 위해 바인딩이 초기화되지 않음")
+class ForyouFragment : Fragment() {
+    private var _binding : FragmentForyouBinding?= null
+    private val binding get () = _binding ?: error("뷰 참조를 위해 바인딩이 초기화되지 않았음")
 
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentSearchBinding.inflate(
+        _binding = FragmentForyouBinding.inflate(
             inflater,
             container,
             false
