+++ app/src/main/java/com/example/seminar5/ui/bottom/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.seminar5.ui
+package com.example.seminar5.ui.bottom
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,11 +6,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.seminar5.R
-import com.example.seminar5.databinding.FragmentMymusicBinding
+import com.example.seminar5.databinding.FragmentHomeBinding
 
-
-class MymusicFragment : Fragment() {
-    private var _binding : FragmentMymusicBinding ?= null
+class HomeFragment : Fragment() {
+    private var _binding : FragmentHomeBinding ?= null
     private val binding get() = _binding ?: error("뷰 참조를 위해 바인딩이 초기화되지 않음")
 
 
@@ -18,19 +17,19 @@ class MymusicFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-       _binding = FragmentMymusicBinding.inflate(
-           inflater,
-           container,
-           false
-       )
+        _binding = FragmentHomeBinding.inflate(
+            inflater,
+            container,
+            false
+        )
         return binding.root
     }
-
 
     override fun onDestroy() {
         super.onDestroy()
         _binding = null
     }
+
 
 
 }
