+++ app/src/main/res/layout/fragment_foryou.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ForyouFragment">
+    tools:context=".ui.bottom.ForyouFragment">
 
 
     <TextView
