+++ app/src/main/res/layout/fragment_mymusic.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout4"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MymusicFragment">
+    tools:context=".ui.bottom.MymusicFragment">
 
     <TextView
         android:id="@+id/textView"
