+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout5"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SearchFragment">
+    tools:context=".ui.bottom.SearchFragment">
 
     <TextView
         android:id="@+id/textView"
