+++ app/src/main/res/values/strings.xml
@@ -14,6 +14,11 @@
     <string name="djchat">DJ차트</string>
     <string name="friend">친구</string>
 
+
+    <!--inhome fragment-->
+    <string name="popularity">실시간 인기</string>
+    <string name="now">최신</string>
+
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
